requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-ckeditor==6.2.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-js-asset==2.0.0
django-uuslug==2.0.0
Pillow==9.0.1
python-slugify==6.1.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
